proto conversion test
